refactor(dwca): isolate definition editor
