+++ specifyweb/frontend/js_src/lib/components/DwcaDefinition/data/coreTermPatterns.ts
