+++ specifyweb/frontend/js_src/lib/components/DwcaDefinition/data/defaultTemplates.ts
